Docs improvements: refactor + content caching + minimap
