+++ src/lsp/gdscript.capabilities.ts
@@ -23,6 +23,7 @@ export class GodotNativeSymbol implements DocumentSymbol {
 	children?: DocumentSymbol[];
 	documentation: string;
 	native_class: string;
+	class_info?: GodotNativeClassInfo;
 }
 
 export interface GodotNativeClassInfo {
